# docs/swagger.yaml
basePath: /v1
definitions:
  errno.Err:
    properties:
      code:
        type: integer
      error:
        type: string
      id:
        type: string
      message:
        type: string
    type: object
...
host: 127.0.0.1:8080
info:
  contact:
    email: lyonsdpy@163.com
    name: 代澎源
    url: https://github.com/lyonsdpy/phoenix
  description: |-
    可用于监控采集,网管信息获取的统一入口.
    后续计划支持prometheus，grpc，etcd.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: phoenix采集agent
  version: 1.0.0
paths:
...
  /snetconf-cap/async:
    post:
      description: SNetconf via ssh获取能力集的异步交互
      parameters:
      - default: 127.0.0.1:830
        description: snetconf请求目标
        in: header
        name: target
        required: true
        type: string
      - description: 用户名
        in: header
        name: username
        required: true
        type: string
      - description: 密码
        in: header
        name: password
        required: true
        type: string
      - default: http://127.0.0.1:8080/v1/common
        description: http回调目标
        in: header
        name: cb_target
        required: true
        type: string
      - description: http回调方法
        enum:
        - POST
        in: header
        name: cb_method
        required: true
        type: string
      produces:
      - text/xml
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/errno.Err'
      tags:
      - SNetconf
  /snetconf-cap/job:
    post:
      description: SNetconf via ssh获取能力集的异步交互
      parameters:
      - default: 127.0.0.1:830
        description: snetconf请求目标
        in: header
        name: target
        required: true
        type: string
      - description: 用户名
        in: header
        name: username
        required: true
        type: string
      - description: 密码
        in: header
        name: password
        required: true
        type: string
      - description: 任务时间间隔
        enum:
        - '@every 5s'
        - '@every 10s'
        - '@hourly'
        - /5 * * * *
        in: header
        name: spec
        required: true
        type: string
      - default: http://127.0.0.1:8080/v1/common
        description: http回调目标
        in: header
        name: cb_target
        required: true
        type: string
      - description: http回调方法
        enum:
        - POST
        in: header
        name: cb_method
        required: true
        type: string
      produces:
      - text/xml
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/errno.Err'
      tags:
      - SNetconf
  /snetconf-cap/sync:
    post:
      description: netconf via ssh的交互,获取能力集
      parameters:
      - default: 127.0.0.1:830
        description: snetconf请求目标
        in: header
        name: target
        required: true
        type: string
      - description: 用户名
        in: header
        name: username
        required: true
        type: string
      - description: 密码
        in: header
        name: password
        required: true
        type: string
      produces:
      - text/xml
      responses:
        "200":
          description: 远端xml
          schema:
            type: string
      tags:
      - SNetconf
  /snetconf/async:
    post:
      consumes:
      - text/xml
      description: SNetconf via ssh的异步交互
      parameters:
      - default: 127.0.0.1:830
        description: snetconf请求目标
        in: header
        name: target
        required: true
        type: string
      - description: 用户名
        in: header
        name: username
        required: true
        type: string
      - description: 密码
        in: header
        name: password
        required: true
        type: string
      - default: http://127.0.0.1:8080/v1/common
        description: http回调目标
        in: header
        name: cb_target
        required: true
        type: string
      - description: http回调方法
        enum:
        - POST
        in: header
        name: cb_method
        required: true
        type: string
      - description: netconf请求xml
        in: body
        name: req_xml
        required: true
        schema:
          type: string
      produces:
      - text/xml
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/errno.Err'
      tags:
      - SNetconf
  /snetconf/job:
    post:
      consumes:
      - text/xml
      description: SNetconf via ssh的异步交互
      parameters:
      - default: 127.0.0.1:830
        description: snetconf请求目标
        in: header
        name: target
        required: true
        type: string
      - description: 用户名
        in: header
        name: username
        required: true
        type: string
      - description: 密码
        in: header
        name: password
        required: true
        type: string
      - description: 任务时间间隔
        enum:
        - '@every 5s'
        - '@every 10s'
        - '@hourly'
        - /5 * * * *
        in: header
        name: spec
        required: true
        type: string
      - default: http://127.0.0.1:8080/v1/common
        description: http回调目标
        in: header
        name: cb_target
        required: true
        type: string
      - description: http回调方法
        enum:
        - POST
        in: header
        name: cb_method
        required: true
        type: string
      - description: netconf请求xml
        in: body
        name: req_xml
        required: true
        schema:
          type: string
      produces:
      - text/xml
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/errno.Err'
      tags:
      - SNetconf
  /snetconf/sync:
    post:
      consumes:
      - text/xml
      description: netconf via ssh的交互
      parameters:
      - default: 127.0.0.1:830
        description: snetconf请求目标
        in: header
        name: target
        required: true
        type: string
      - description: 用户名
        in: header
        name: username
        required: true
        type: string
      - description: 密码
        in: header
        name: password
        required: true
        type: string
      - description: netconf请求xml
        in: body
        name: req_xml
        required: true
        schema:
          type: string
      produces:
      - text/xml
      responses:
        "200":
          description: 远端xml
          schema:
            type: string
      tags:
      - SNetconf
...
swagger: "2.0"